Split chatroom.py
