+++ index.py
@@ -3,7 +3,6 @@ from flask_sqlalchemy import SQLAlchemy
 from config import BaseConfig
 from flask_bcrypt import Bcrypt
 from flask_socketio import SocketIO
-from flask_cors import CORS
 
 app = Flask(__name__, static_folder="./static/dist", template_folder="./static")
 
@@ -11,6 +10,5 @@ app = Flask(__name__, static_folder="./static/dist", template_folder="./static")
 
 app.config.from_object(BaseConfig)
 socketio = SocketIO(app, cors_allowed_origins='*')
-CORS(app)
 db = SQLAlchemy(app)
 bcrypt = Bcrypt(app)
